[TEST]:Move tests folder to src and renamed to test
